Fix a crash when InitMiners has not completed
If `InitMiners` has not completed (because LXRHash is building the 1GB
table file) and the `ReportHashRate` timer is triggered, the
`SendCommand` call dereferences a null pointer.  (It might be a good
idea to add a check in `SendCommand`.)  This change defers starting the
`ReportHashRate` goroutine until the end of `InitMiners`.