Add mining service
This change, which has only been tested on Windows 10, adds a Windows
service (eventually also a *NIX/macOS daemon.)  The newly-added service
command can start, stop, install and uninstall the service on Windows,
provided the account running the command has sufficient privileges.
This service has a JSON-RPC server which listens on a named pipe.  The
change also adds the rpc and feed commands, which can control and query
the state of the miner service.